refactor(sage_sms): change backend dir to backends
