--- sage_sms/utils.py
@@ -1,6 +1,0 @@
-import toml
-
-def load_settings(toml_file):
-    with open(toml_file, 'r') as f:
-        config = toml.load(f)
-    return config
